Cleanup of no-unnecessary-type-assertion
* Replace references to cast with the more correct type assertion
* Fix formatting
* Improve metadata
