--- test/rules/no-unnecessary-cast/test.ts.lint
@@ -1,88 +1,0 @@
-const nonNullStringLiteral: 'test';
-const nonNullString: string;
-const nullableString: string|undefined;
-
-// non-null
-let a = nonNullStringLiteral!;
-        ~~~~~~~~~~~~~~~~~~~~~  [This cast is unnecessary.]
-let b = nonNullString!;
-        ~~~~~~~~~~~~~~  [This cast is unnecessary.]
-let c = nullableString!;
-
-// as
-let d = nonNullStringLiteral as string;
-let e = nonNullString as string;
-        ~~~~~~~~~~~~~~~~~~~~~~~  [This cast is unnecessary.]
-let f = nullableString as string;
-
-// type assertion
-let g = <string>nonNullStringLiteral;
-let h = <string>nonNullString;
-        ~~~~~~~~~~~~~~~~~~~~~  [This cast is unnecessary.]
-let i = <string>nullableString;
-
-// complex inner expression
-let j = (nonNullString + nonNullStringLiteral)!;
-        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~  [This cast is unnecessary.]
-let k = (nonNullString + nonNullStringLiteral) as string;
-        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~  [This cast is unnecessary.]
-let l = <string>(nonNullString + nonNullStringLiteral);
-        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~  [This cast is unnecessary.]
-let m = nonNullString.trim()!;
-        ~~~~~~~~~~~~~~~~~~~~~  [This cast is unnecessary.]
-let n = nonNullString.trim() as string;
-        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~  [This cast is unnecessary.]
-let o = <string>nonNullString.trim();
-        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~  [This cast is unnecessary.]
-let p = nonNullString!.trim();
-        ~~~~~~~~~~~~~~  [This cast is unnecessary.]
-
-// custom types
-interface Iface1 {
-  prop: string;
-}
-interface Iface2 {
-  prop: string;
-}
-
-const value1: Iface1 = {prop: 'test'};
-const value2: Iface2 = {prop: 'test'};
-
-let q = <Iface1>value1;
-        ~~~~~~~~~~~~~~  [This cast is unnecessary.]
-let r = <Iface2>value1;
-let s = value2 as Iface2;
-        ~~~~~~~~~~~~~~~~  [This cast is unnecessary.]
-let t = value2 as Iface1;
-
-interface TypeA {
-  kind: 'a';
-}
-interface TypeB {
-  kind: 'b';
-}
-
-function isB(x: TypeA|TypeB): x is TypeB {
-  return true;
-}
-
-function func(aOrB: TypeA|TypeB) {
-  let u = aOrB as TypeA;
-  let v = <TypeB>aOrB;
-
-  if (aOrB.kind === 'a') {
-    let w = aOrB as TypeA;
-            ~~~~~~~~~~~~~  [This cast is unnecessary.]
-  } else {
-    let x = <TypeB>aOrB;
-            ~~~~~~~~~~~  [This cast is unnecessary.]
-  }
-
-  if (isB(aOrB)) {
-    let y = aOrB as TypeB;
-            ~~~~~~~~~~~~~  [This cast is unnecessary.]
-  } else {
-    let z = <TypeA>aOrB;
-            ~~~~~~~~~~~  [This cast is unnecessary.]
-  }
-}
