+++ test/rules/no-unnecessary-type-assertion/test.ts.fix
@@ -28,10 +28,10 @@ let p = nonNullString.trim();
 
 // custom types
 interface Iface1 {
-  prop: string;
+    prop: string;
 }
 interface Iface2 {
-  prop: string;
+    prop: string;
 }
 
 const value1: Iface1 = {prop: 'test'};
@@ -43,29 +43,29 @@ let s = value2;
 let t = value2 as Iface1;
 
 interface TypeA {
-  kind: 'a';
+    kind: 'a';
 }
 interface TypeB {
-  kind: 'b';
+    kind: 'b';
 }
 
 function isB(x: TypeA|TypeB): x is TypeB {
-  return true;
+    return true;
 }
 
 function func(aOrB: TypeA|TypeB) {
-  let u = aOrB as TypeA;
-  let v = <TypeB>aOrB;
+    let u = aOrB as TypeA;
+    let v = <TypeB>aOrB;
 
-  if (aOrB.kind === 'a') {
-    let w = aOrB;
-  } else {
-    let x = aOrB;
-  }
+    if (aOrB.kind === 'a') {
+        let w = aOrB;
+    } else {
+        let x = aOrB;
+    }
 
-  if (isB(aOrB)) {
-    let y = aOrB;
-  } else {
-    let z = aOrB;
-  }
+    if (isB(aOrB)) {
+        let y = aOrB;
+    } else {
+        let z = aOrB;
+    }
 }
