+++ test/rules/no-unnecessary-type-assertion/test.ts.lint
@@ -1,0 +1,88 @@
+const nonNullStringLiteral: 'test';
+const nonNullString: string;
+const nullableString: string|undefined;
+
+// non-null
+let a = nonNullStringLiteral!;
+        ~~~~~~~~~~~~~~~~~~~~~  [This assertion is unnecessary since it does not change the type of the expression.]
+let b = nonNullString!;
+        ~~~~~~~~~~~~~~  [This assertion is unnecessary since it does not change the type of the expression.]
+let c = nullableString!;
+
+// as
+let d = nonNullStringLiteral as string;
+let e = nonNullString as string;
+        ~~~~~~~~~~~~~~~~~~~~~~~  [This assertion is unnecessary since it does not change the type of the expression.]
+let f = nullableString as string;
+
+// type assertion
+let g = <string>nonNullStringLiteral;
+let h = <string>nonNullString;
+        ~~~~~~~~~~~~~~~~~~~~~  [This assertion is unnecessary since it does not change the type of the expression.]
+let i = <string>nullableString;
+
+// complex inner expression
+let j = (nonNullString + nonNullStringLiteral)!;
+        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~  [This assertion is unnecessary since it does not change the type of the expression.]
+let k = (nonNullString + nonNullStringLiteral) as string;
+        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~  [This assertion is unnecessary since it does not change the type of the expression.]
+let l = <string>(nonNullString + nonNullStringLiteral);
+        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~  [This assertion is unnecessary since it does not change the type of the expression.]
+let m = nonNullString.trim()!;
+        ~~~~~~~~~~~~~~~~~~~~~  [This assertion is unnecessary since it does not change the type of the expression.]
+let n = nonNullString.trim() as string;
+        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~  [This assertion is unnecessary since it does not change the type of the expression.]
+let o = <string>nonNullString.trim();
+        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~  [This assertion is unnecessary since it does not change the type of the expression.]
+let p = nonNullString!.trim();
+        ~~~~~~~~~~~~~~  [This assertion is unnecessary since it does not change the type of the expression.]
+
+// custom types
+interface Iface1 {
+    prop: string;
+}
+interface Iface2 {
+    prop: string;
+}
+
+const value1: Iface1 = {prop: 'test'};
+const value2: Iface2 = {prop: 'test'};
+
+let q = <Iface1>value1;
+        ~~~~~~~~~~~~~~  [This assertion is unnecessary since it does not change the type of the expression.]
+let r = <Iface2>value1;
+let s = value2 as Iface2;
+        ~~~~~~~~~~~~~~~~  [This assertion is unnecessary since it does not change the type of the expression.]
+let t = value2 as Iface1;
+
+interface TypeA {
+    kind: 'a';
+}
+interface TypeB {
+    kind: 'b';
+}
+
+function isB(x: TypeA|TypeB): x is TypeB {
+    return true;
+}
+
+function func(aOrB: TypeA|TypeB) {
+    let u = aOrB as TypeA;
+    let v = <TypeB>aOrB;
+
+    if (aOrB.kind === 'a') {
+        let w = aOrB as TypeA;
+                ~~~~~~~~~~~~~  [This assertion is unnecessary since it does not change the type of the expression.]
+    } else {
+        let x = <TypeB>aOrB;
+                ~~~~~~~~~~~  [This assertion is unnecessary since it does not change the type of the expression.]
+    }
+
+    if (isB(aOrB)) {
+        let y = aOrB as TypeB;
+                ~~~~~~~~~~~~~  [This assertion is unnecessary since it does not change the type of the expression.]
+    } else {
+        let z = <TypeA>aOrB;
+                ~~~~~~~~~~~  [This assertion is unnecessary since it does not change the type of the expression.]
+    }
+}
