+++ test/rules/no-unnecessary-type-assertion/tsconfig.json
