+++ test/rules/no-unnecessary-type-assertion/tslint.json
@@ -3,6 +3,6 @@
     "typeCheck": true
   },
   "rules": {
-    "no-unnecessary-cast": true
+    "no-unnecessary-type-assertion": true
   }
 }
